drf/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
mysqlclient==2.2.4
gunicorn==21.2.0
django-cors-headers==4.4.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1

